Mention the wireshark-common prompt for non-Pi installations.
That's not shown with Pi images since we're preseeding answers there.

That wasn't shown during ViRogue preparations since the dependencies
were different.

With the new dependencies in pirogue-base 2.0.4, this one prompt is
back.

Both v4/v6 prompts from iptables-persistent are a thing from the past
though, as it's no longer listed in dependencies.